Extracted DataViewer and renamed modal
